perf: keep selector perps-worth machinery out of bg startup graph
- Move perps net-worth gating/resolution from ServiceAccountSelector into
  utils/accountSelectorPerpsWorth, loaded via dynamic import on first use
- Lazy-import mergeBatchCreateCustomNetworks at its call site
- Keeps the native background startup graph within the size budget: the
  base branch has only ~0.6 KB headroom, so selector-open-only code must
  stay behind lazy segments
