+++ packages/kit-bg/src/services/ServiceBatchCreateAccount/ServiceBatchCreateAccount.ts
@@ -66,7 +66,6 @@ import { buildDefaultAddAccountNetworks } from '../ServiceAccount/defaultNetwork
 import ServiceBase from '../ServiceBase';
 import { HardwareAllNetworkGetAddressResponse } from '../ServiceHardware/HardwareAllNetworkGetAddressResponse';
 
-import { mergeBatchCreateCustomNetworks } from './batchCreateCustomNetworks';
 import { normalizeAllNetworkInstallCancelErrors } from './thirdPartyAllNetworkErrors';
 import {
   type IThirdPartyAllNetworkAddressParams,
@@ -431,6 +430,11 @@ class ServiceBatchCreateAccount extends ServiceBase {
         ];
 
         if (payload.params.customNetworks) {
+          // Dynamic import keeps the merge helper out of the native
+          // background startup graph (Startup Graph Budget check).
+          const { mergeBatchCreateCustomNetworks } = await import(
+            './batchCreateCustomNetworks'
+          );
           customNetworks = mergeBatchCreateCustomNetworks({
             defaultNetworks: customNetworks,
             customNetworks: payload.params.customNetworks,
